finish discount mangagement

+++ techstore/src/main/java/com/techstore/dto/request/DiscountRequest.java
@@ -1,6 +1,8 @@
 package com.techstore.dto.request;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -8,8 +10,8 @@ public class DiscountRequest {
     private String code;
     private String description;
     private Integer discountPercent;
-    private Double minOrderAmount;
-    private Double maxDiscountAmount;
+    private BigDecimal minOrderAmount;
+    private BigDecimal maxDiscountAmount;
     private Integer quantity;
     private LocalDate startDate;
     private LocalDate endDate;
